fix: add a base fly app configuration
`fly launch` defaults to apps stopping after a period of
inactivity. Since a persistent volume is also not configured, this
would result in historical data being lost on every pause.

This change adds a fly.toml that disables auto-stop for the app that
will be launched.

Additionally, a persistent volume is also defined, and the readme is
updated to indicate that it needs to be explicitly created.

In theory the `fly launch` command should pick up the config from the
fly.toml file. In practice this doesn't seem to happen.